After a few sleepless nights. I gained almost certain to fight against a specific PHP bug to 7 at least until the 7.0.7 release.
At the time of serialization, the entry into array Election->_Candidates become references. Unknown problem with PHP 5.6, but not reproducible in simpler tests under PHP 7. That sounds really perverse.

This patch adds a corrective action at the wakeup time.

It will try to reverse this patch with the release of the next version of PHP 7.0.x. The last changelogs already talking a lot of particular issues with references.