Registered Vote object can be edited in certain circonstances, when using an external data handler.
This remains an unpredictable and unrecommended feature. And backed by a complex code.
But the objective of this commit is to reduce the zones of differentiation between the standard behavior (without dataHandler) and the behavior with dataHandler.
However, if you manage millions of votes, avoid editing or deleting them in the same Condorcet election.